add 'snap to' first product placed in the container.
Added a 'user' StorageSettings object that the user interacts with (thingfilter, priority, copy/paste buttons), and that propagates changes to the now hidden 'real' StorageSettings object used by the game.

On adding something to the container, the 'real' filter is limited to just that def, forbidding all other input. When emptied (or current contents are no longer allowd), the real filter is reset.