add ExposeData logic
Added userSettings to ExposeData. I realised we also need to re-apply the callback on the userSettings.filter, so I added a reflection wrapper to set it directly, rather than faffing about with creating new filters.

Also added some migration logic to add userSettings if it was null (e.g. on updating the mod in old saves).